feat: rewrite log collector outside of `Actions`
