+++ uplink/src/collector/mod.rs
@@ -1,3 +1,4 @@
+pub mod logging;
 pub mod simulator;
 pub mod systemstats;
 pub mod tcpjson;
